refactor(blockifier,starknet_api): migrate GasVector from blockifier to starknet_api
