+++ crates/blockifier/src/blockifier/block.rs
@@ -1,5 +1,12 @@
 use log::warn;
-use starknet_api::block::{BlockHash, BlockNumber, BlockTimestamp, GasPrice, NonzeroGasPrice};
+use starknet_api::block::{
+    BlockHash,
+    BlockNumber,
+    BlockTimestamp,
+    GasPrice,
+    GasPricesForFeeType,
+    NonzeroGasPrice,
+};
 use starknet_api::core::ContractAddress;
 use starknet_api::state::StorageKey;
 use starknet_types_core::felt::Felt;
@@ -29,13 +36,6 @@ pub struct BlockInfo {
 pub struct GasPrices {
     eth_gas_prices: GasPricesForFeeType,  // In wei.
     strk_gas_prices: GasPricesForFeeType, // In fri.
-}
-
-#[derive(Clone, Debug)]
-pub struct GasPricesForFeeType {
-    pub l1_gas_price: NonzeroGasPrice,
-    pub l1_data_gas_price: NonzeroGasPrice,
-    pub l2_gas_price: NonzeroGasPrice,
 }
 
 impl GasPrices {
