+++ crates/blockifier/src/fee/receipt.rs
@@ -1,12 +1,12 @@
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use starknet_api::core::ContractAddress;
+use starknet_api::execution_resources::GasVector;
 use starknet_api::transaction::Fee;
 
 use crate::context::TransactionContext;
 use crate::execution::call_info::ExecutionSummary;
 use crate::fee::resources::{
     ComputationResources,
-    GasVector,
     StarknetResources,
     StateResources,
     TransactionResources,
