+++ crates/starknet_api/src/block.rs
@@ -281,6 +281,13 @@ impl TryFrom<GasPrice> for NonzeroGasPrice {
     }
 }
 
+#[derive(Clone, Debug)]
+pub struct GasPricesForFeeType {
+    pub l1_gas_price: NonzeroGasPrice,
+    pub l1_data_gas_price: NonzeroGasPrice,
+    pub l2_gas_price: NonzeroGasPrice,
+}
+
 /// The timestamp of a [Block](`crate::block::Block`).
 #[derive(
     Debug, Default, Copy, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord,
